rename database/ <- data/
